custom-elements/custom-element-lifecycle/types-of-callbacks/attribute-changed-callback-test-002.html

<!DOCTYPE html>
<html>
<head>
<title>Test attribute changing to check attributeChanged callback of a custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="attributeChanged callback must be enqueued whenever custom element's attribute is added, changed or removed">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.register('x-a-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    customElement.setAttribute('class', 'someClass');
    customElement.setAttribute('class', 'someClass2');
    assert_equals(callbackCalledCounter, 2, 'Callback attributeChanged was called ' +
            'wrong number of times');

    customElement.classList.add('someClass3');
    assert_equals(callbackCalledCounter, 3, 'Callback attributeChanged was called ' +
            'wrong number of times after classList.add');

    customElement.id = 'someId';
    assert_equals(callbackCalledCounter, 4, 'Callback attributeChanged was called ' +
            'wrong number of times after changing attribute as property');
}, 'Test attributeChanged callback of the custom element created in JavaScript. ' +
    'Test that callback is called when attribute is changed');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;
    var argument1 = 'arg1';
    var argument2 = 'arg2';
    var argument3 = 'arg3';

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        callbackCalledCounter++;
        argument1 = arg1;
        argument2 = arg2;
        argument3 = arg3;
    }
    doc.register('x-a-2', {prototype: proto});
    doc.body.innerHTML = '<x-a-2 id="xa2" class="oldValue"></x-a-2>';
    var customElement = doc.querySelector('#xa2');

    customElement.setAttribute('class', 'newValue');
    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged was called ' +
            'wrong number of times');
    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
        'attributeChanged callback');
    assert_equals(argument2, 'oldValue', 'Wrong value of the second argument of ' +
        'attributeChanged callback');
    assert_equals(argument3, 'newValue', 'Wrong value of the third argument of ' +
        'attributeChanged callback');

    customElement.classList.add('newestValue');
    assert_equals(callbackCalledCounter, 2, 'Callback attributeChanged was called ' +
            'wrong number of times after changing attribute by classList.add');
    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
        'attributeChanged callback after changing attribute by classList.add');
    assert_equals(argument2, 'newValue', 'Wrong value of the second argument of ' +
        'attributeChanged callback after changing attribute by classList.add');
    assert_equals(argument3, 'newValue newestValue', 'Wrong value of the third argument of ' +
        'attributeChanged callback after changing attribute by classList.add');

    customElement.className = 'lastValue';
    assert_equals(callbackCalledCounter, 3, 'Callback attributeChanged was called ' +
            'wrong number of times after changing attribute as a property');
    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
        'attributeChanged callback after changing attribute by classList.add');
    assert_equals(argument2, 'newValue newestValue', 'Wrong value of the second argument of ' +
        'attributeChanged callback after changing attribute by classList.add');
    assert_equals(argument3, 'lastValue', 'Wrong value of the third argument of ' +
        'attributeChanged callback after changing attribute by classList.add');

}, 'Test attributeChanged callback of the custom element instantiating in HTML. ' +
    'Test callback arguments when attribute is changed');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;
    var argument1 = 'arg1';
    var argument2 = 'arg2';
    var argument3 = 'arg3';

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function (arg1, arg2, arg3) {
        callbackCalledCounter++;
        argument1 = arg1;
        argument2 = arg2;
        argument3 = arg3;
    }
    doc.body.innerHTML = '<x-a-3 id="xa3" class="oldValue"></x-a-3>';
    var customElement = doc.querySelector('#xa3');
    customElement.setAttribute('class', 'newValue');
    customElement.setAttribute('class', 'newestValue');
    doc.register('x-a-3', {prototype: proto});
    customElement.setAttribute('class', 'rightValue');

    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged was called ' +
            'wrong number of times');
    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
            'attributeChanged callback');
    assert_equals(argument2, 'newestValue', 'Wrong value of the second argument of ' +
            'attributeChanged callback');
    assert_equals(argument3, 'rightValue', 'Wrong value of the third argument of ' +
            'attributeChanged callback');
}, 'Test attributeChanged callback of the custom element instantiating in HTML. ' +
    'Test attribute changing of unregistered element');


testInIFrame('../../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-element', {prototype: proto});
    var customElement = doc.querySelector('#xelement');
    customElement.setAttribute('class', 'firstValue');
    customElement.setAttribute('class', 'secondValue');
    assert_equals(callbackCalledCounter, 2, 'Callback attributeChanged was called ' +
            'wrong number of times');

    customElement.classList.add('someClass3');
    assert_equals(callbackCalledCounter, 3, 'Callback attributeChanged was called ' +
            'wrong number of times after classList.add');

    customElement.id = 'someId';
    assert_equals(callbackCalledCounter, 4, 'Callback attributeChanged was called ' +
            'wrong number of times after changing attribute as property');

}, 'Test attributeChanged callback for document that has browsing context. ' +
    'Test that callback is called when attribute is changed');


testInIFrame('../../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var argument1 = 'arg1';
    var argument2 = 'arg2';
    var argument3 = 'arg3';

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function (arg1, arg2, arg3) {
        argument1 = arg1;
        argument2 = arg2;
        argument3 = arg3;
    }
    doc.register('x-element', {prototype: proto});
    var customElement = doc.querySelector('#xelement');
    customElement.setAttribute('class', 'firstValue');
    customElement.setAttribute('class', 'secondValue');
    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
        'attributeChanged callback');
    assert_equals(argument2, 'firstValue', 'Wrong value of the second argument of ' +
        'attributeChanged callback');
    assert_equals(argument3, 'secondValue', 'Wrong value of the third argument of ' +
        'attributeChanged callback');

    customElement.classList.add('newestValue');
    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
        'attributeChanged callback after changing attribute by classList.add');
    assert_equals(argument2, 'secondValue', 'Wrong value of the second argument of ' +
        'attributeChanged callback after changing attribute by classList.add');
    assert_equals(argument3, 'secondValue newestValue', 'Wrong value of the third argument of ' +
        'attributeChanged callback after changing attribute by classList.add');

    customElement.className = 'lastValue';
    assert_equals(argument1, 'class', 'Wrong value of the first argument of ' +
        'attributeChanged callback after changing attribute as a property');
    assert_equals(argument2, 'secondValue newestValue', 'Wrong value of the second argument of ' +
        'attributeChanged callback after changing attribute as a property');
    assert_equals(argument3, 'lastValue', 'Wrong value of the third argument of ' +
        'attributeChanged callback after changing attribute as a property');

}, 'Test attributeChanged callback for document that has browsing context. ' +
    'Test callback attributes when attribute is changed');


testInIFrame('../../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function () {
        callbackCalledCounter++;
    }
    var customElement = doc.querySelector('#xelement');
    customElement.setAttribute('class', 'firstValue');
    customElement.setAttribute('class', 'secondValue');
    doc.register('x-element', {prototype: proto});
    customElement.setAttribute('class', 'thirdValue');
    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged was called ' +
            'wrong number of times');
}, 'Test attributeChanged callback for document that has browsing context. ' +
    'Test that callback is called when attribute is changed. Element is unregistered ' +
    'at first');

</script>
</body>
</html>
